Add files via upload
This dashboard shows that Drama is the most popular genre on Prime Video, followed by Comedy. There are also a significant number of shows in the categories of Drama, Suspense and Comedy, Drama. The majority of the shows on Prime Video are from Africa. There are more TV shows on Prime Video than movies.
In this we generated the interactive Dashboard.